Feat/presentation tesis 2025

+++ app/src/main/java/com/durand/dogedex/data/request/oficial/RegisterCanRequest.kt
@@ -32,5 +32,5 @@ data class RegisterCanRequest(
     @SerializedName("foto")
     var foto: String,
     @SerializedName("idUsuario")
-    var idUsuario: Int,
+    var idUsuario: Long,
 )
